refactor(worker): give Temporal workflows and activities a real structure
The worker had accumulated a 2,236-line index.ts holding process lifecycle,
Gmail sync, history replay, Memory extraction, and the Activity adapter in one
file, while packages/workflows kept its contract and its single workflow body
flat at the package root. Neither layout left an obvious place to add a
workflow, and the sandbox boundary that makes workflow code deterministic was
not visible in the tree.

packages/workflows now separates the three things it owns:

  task-queues.ts             lane names, free of any Temporal SDK import so
                             packages/database can route without the dependency
  contracts/workflow-step.ts wire types shared by the sandbox and the worker
  workflows/workflow-step.ts the workflow body

The contracts/ and workflows/ pairing is the shape each further workflow
follows.

apps/worker splits along its two real responsibilities. temporal/ holds the
runtime: environment parsing, Cloud configuration and queue naming, the
TemporalRuntime class, and the sandbox bundle entry. activities/ holds the
executable work, grouped by the integration that owns it -- gmail, label, and
memory -- with registry.ts as the single place a step type becomes a handler.
index.ts keeps only process lifecycle and is now 185 lines. The activity
modules form an acyclic graph: connection to messages to drafts to watch to
history to sync.

GMAIL_CONTENT_CONCURRENCY moves out of the Temporal runtime, where it did not
belong, into activities/gmail/concurrency.ts. It bounds provider fetches, not
Activity execution.

Test globs for the two restructured packages become recursive so colocated
tests keep running from their new directories.

Behavior is unchanged; this is a move. Verified with pnpm typecheck across all
ten projects, pnpm test at 300 tests and zero failures with the worker's 38
matching its pre-refactor baseline, pnpm lint, pnpm build, and a direct
bundleWorkflowCode run against the relocated sandbox entry, which compiles at
1.50 MB. Nothing was exercised against Temporal Cloud, Gmail, or PostgreSQL, so
worker startup and the command dispatch loop remain unverified at runtime.
