+++ apps/worker/src/activities/memory/feedback.ts
@@ -1,0 +1,123 @@
+/**
+ * Turns user feedback on assistant output into stored Memory candidates.
+ */
+import {
+  AiConfigurationError,
+  extractFeedbackMemories,
+  isAiConfigured,
+  type FeedbackMemoryCandidate,
+} from "@invook/ai";
+import {
+  DRAFT_FEEDBACK_VERSION,
+  getDraftFeedbackSamples,
+  getUserAuthoredMemories,
+  getWorkerAccount,
+  markDraftFeedbackAnalyzed,
+  saveExtractedMemories,
+  type MemoryType,
+  type WorkflowStepJob,
+} from "@invook/database";
+
+import { feedbackBatchSize } from "../configuration";
+import { normalizedEmails } from "../gmail/messages";
+
+export async function runMemoryFeedback(job: WorkflowStepJob) {
+  if (!job.accountId) throw new Error("The feedback job has no connected account.");
+  const account = await getWorkerAccount(job.accountId);
+  if (!account) throw new Error("The connected Gmail account was not found.");
+
+  const samples = await getDraftFeedbackSamples(
+    account.id,
+    DRAFT_FEEDBACK_VERSION,
+    feedbackBatchSize,
+  );
+  if (samples.length < 3) {
+    return { status: "waiting_for_repetition", sampleCount: samples.length };
+  }
+  if (!isAiConfigured()) throw new AiConfigurationError();
+
+  const protectedMemories = await getUserAuthoredMemories(account.id);
+  const analysis = await extractFeedbackMemories({
+    protectedMemories,
+    samples: samples.flatMap((sample) =>
+      sample.generatedText
+        ? [
+            {
+              id: sample.id,
+              subject: sample.subject,
+              contactEmails: normalizedEmails(sample.participants, account.email),
+              generatedText: sample.generatedText,
+              editedText: sample.editedText,
+            },
+          ]
+        : [],
+    ),
+  });
+
+  const samplesById = new Map(samples.map((sample) => [sample.id, sample]));
+  const validMemories: FeedbackMemoryCandidate[] = [];
+  for (const memory of analysis.memories) {
+    const evidenceDraftIds = Array.from(new Set(memory.evidenceDraftIds));
+    if (
+      evidenceDraftIds.length < 3 ||
+      evidenceDraftIds.some((id) => !samplesById.has(id))
+    ) {
+      continue;
+    }
+
+    if (memory.type === "contact") {
+      const contactEmail = memory.contactEmail?.trim().toLowerCase();
+      if (!contactEmail) continue;
+      const repeatedForContact = evidenceDraftIds.every((id) => {
+        const sample = samplesById.get(id);
+        return Boolean(
+          sample && normalizedEmails(sample.participants, account.email).includes(contactEmail),
+        );
+      });
+      if (!repeatedForContact) continue;
+      validMemories.push({ ...memory, contactEmail, evidenceDraftIds });
+      continue;
+    }
+
+    if (memory.type === "preference") {
+      const contacts = new Set(
+        evidenceDraftIds.flatMap((id) => {
+          const sample = samplesById.get(id);
+          return sample ? normalizedEmails(sample.participants, account.email) : [];
+        }),
+      );
+      if (contacts.size < 3) continue;
+    }
+    validMemories.push({ ...memory, contactEmail: null, evidenceDraftIds });
+  }
+
+  const savedCount = await saveExtractedMemories({
+    userId: account.userId,
+    accountId: account.id,
+    source: "feedback",
+    modelId: analysis.modelId,
+    memories: validMemories,
+  });
+
+  const signalsByDraft = new Map<
+    string,
+    Array<{ type: MemoryType; statement: string }>
+  >();
+  for (const memory of validMemories) {
+    for (const draftId of memory.evidenceDraftIds) {
+      const signals = signalsByDraft.get(draftId) ?? [];
+      signals.push({ type: memory.type, statement: memory.statement });
+      signalsByDraft.set(draftId, signals);
+    }
+  }
+  await markDraftFeedbackAnalyzed({
+    draftIds: samples.map((sample) => sample.id),
+    signalsByDraft,
+  });
+
+  return {
+    status: "complete",
+    sampleCount: samples.length,
+    memoryCount: savedCount,
+  };
+}
